+++ examples/llm_ptq/hf_ptq.py
@@ -575,7 +575,10 @@ def pre_quantize(
     ][0:1]
 
     # Generate preview before quantization
-    if is_nemotron_vl_model and tokenizer is not None:
+    if model_type == "deepseek":
+        # DeepSeek generation may go OOM, so we skip it
+        generated_ids_before_ptq = None
+    elif is_nemotron_vl_model and tokenizer is not None:
         generated_ids_before_ptq = run_nemotron_vl_preview(
             full_model,
             tokenizer,
@@ -618,7 +621,9 @@ def post_quantize(
     # Run some samples
     torch.cuda.empty_cache()
     generated_ids_after_ptq = None
-    if model_type != "llama4" and not is_nemotron_vl_model:
+    if generated_ids_before_ptq is None:
+        pass
+    elif model_type != "llama4" and not is_nemotron_vl_model:
         # Our fake quantizer may not be fully compatible with torch.compile.
         generated_ids_after_ptq = full_model.generate(preview_input_ids, max_new_tokens=100)
     elif is_nemotron_vl_model and tokenizer is not None: